Added coeff parameter to solveVelocity function.
Changes in the solveVelocity function of lagrangian_hydro.py.

coeff parameter is A * dt / m_half.

This was already defined in the loop over interior cell edges, but
not implemented.

Changes made added extra parameters coeff_L and coeff_R for
boundary velocity computations for a given pressure boundary
condition, and implemented all three parameters in the solve for
the velocities.