Implemented time-stepping.
Added inputs for number of spatial cells and system size.

Added time and time step number member variables to radpydro so that 
current time can be tracked in addition to the time step number

In computeTimeStep, the derivative of E was corrected (simple sign 
error).

Function added so that the solution is plotted. Currently this plots 
everything on a subplot.

Energy conservation function needs to be updated to work over several 
time-steps. Conservation is only correct for the first time step.

I suggest we should add attributes to fields for kinetic energy, 
internal energy, radiation energy, and radiation energy leakage for easy 
computations without having to store everything.